requirements.txt
blake3==0.3.1
blake512==1.0.0
gostcrypto==1.2.5
groestlcoin-hash2==1.1.1
pysha3==1.0.2
pyskein==1.0
Whirlpool==1.0.0
